+++ Frontend/Vue/src/views/GoalViews/setFitnessGoal.vue
@@ -21,7 +21,8 @@ export default {
         description: '',
         currentProgress: '',
         targetDate: '',
-        notes: ''
+        notes: '',
+        is_done: false
       },
       errorMessage: ''
     }